Create the reader input file set for streamlining the downstream MPI distribution (#1109)
* Got collection of associated ranks working.

* Collection of location indices from all ranks onto rank 0 is working.

* WIP: have loop for building each input file constructed. No files yet, just the proper
loop structure in place.

* WIP: have the beginnings of the input file set. Files are being created, but not complete yet.

* WIP: dateTime, latitude and longitude (varibles that have been prepared by initialization
prior to the creation of the new input file set) are getting created and data transferred
correctly.

* Refactored code that copies the group heierarchical structure in order to remove duplicated
code.

* WIP: the proper group hierarchical structures including dimensions and variables are
being created in the input file set. All that is left is the varible data transfer.

* WIP: variable data transfer is working except for setting the proper missing values.

* Input file set is getting fully created. Need to test some more.

* Removed some unnecessary debug statements. Added setting of stringMissingValue_ data
member for ReaderSinglePool object.

* Placed functions that were moved to satisfy declaration order back in their original spot,
and then placed forward declarations at the top of the file. This is the start of a
framework for making all of the routines not declared in the ReaderPoolUtils.h file
truly private (static).

* Added "const" qualifier to a number of variables that only get assigned once.
Changed the variable creation and data transfer debug statements from the oops::Log::debug()
channel to the oops::Log::trace() channel since these are really tracing information.